+++ integration-libs/opf/base/core/services/index.ts
@@ -6,3 +6,4 @@
 
 export * from './opf-endpoints.service';
 export * from './opf-payment-error-handler.service';
+export * from './opf-payment-hosted-fields.service';
